Fix int-long mismatch in device discovery

cregit: 5cbdba588a14e05fd282b5850a3e7b66fd9b2b65 https://github.com/qt/qtbase/commit/5cbdba588a14e05fd282b5850a3e7b66fd9b2b65

+++ src/platformsupport/devicediscovery/qdevicediscovery_static.cpp
@@ -69,7 +69,7 @@ DECL|macro|LONG_BITS
 define|#
 directive|define
 name|LONG_BITS
-value|(sizeof(int) * 8 )
+value|(sizeof(long) * 8 )
 end_define
 begin_define
 DECL|macro|LONG_FIELD_SIZE
